+++ packages/stream_chat/lib/src/core/api/channel_api.dart
@@ -292,6 +292,7 @@ class ChannelApi {
   ) async {
     final response = await _client.post(
       '${_getChannelUrl(channelId, channelType)}/show',
+      data: {},
     );
     return EmptyResponse.fromJson(response.data);
   }