+++ packages/stream_chat/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Upcoming
 
 - Bumped dependencies
-    - async: ^2.5.0 -> ^2.8.1
+    - async: ^2.5.0 -> ^2.6.1
     - equatable: ^2.0.0 -> ^2.0.3
     - freezed_annotation: ^0.14.0 -> ^0.14.3
     - json_annotation: ^4.0.1 -> ^4.1.0