+++ packages/stream_chat/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Upcoming
+
+🐞 Fixed
+
+- Fix `channel.show` not working because of null body
+
 ## 2.2.0
 
 🐞 Fixed