+++ packages/stream_chat/CHANGELOG.md
@@ -3,6 +3,7 @@
 🐞 Fixed
 
 - Fixed unread indicator not updating correctly
+- Fix `channel.show` not working because of null body
 
 ## 2.2.0