+++ packages/stream_chat/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 🐞 Fixed
 
+- Fixed unread indicator not updating correctly
 - Fix `channel.show` not working because of null body
 
 ## 2.2.0